requirements.txt
setuptools>=65.0.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.37.0
plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
supabase>=2.0.0
PyJWT>=2.8.0
extra-streamlit-components>=0.1.60
